requirements.txt
# Core dependencies
psycopg2-binary==2.9.10
pandas==2.3.1
numpy==2.3.1

# Document parsing
pdfplumber==0.11.7
pdfminer-six==20250506
pypdfium2==4.30.1

# Utilities
tenacity==9.1.2
colorama==0.4.6
python-dotenv==1.0.0

# Testing
pytest==8.4.1

# Additional dependencies from lock file
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
iniconfig==2.1.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pycparser==2.22
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2